Move hash combining to common::hash
